Fire feature should have it's own module
